Move things into library files
Move audio init into setup.py
Move midi functions into midi.py
Move menu functions into menu.py
Move state to state.py
Move FlashyState to flashy.py
Refactor menuing to reduce use of global state
