+++ Software/Testing/state.py
@@ -1,0 +1,16 @@
+class State(object):
+    def __init__(self):
+        pass
+
+    @property
+    def name(self):
+        return ""
+
+    def enter(self, machine):
+        pass
+
+    def exit(self, machine):
+        pass
+
+    def update(self, machine):
+        pass
